+++ packages/sources/finage/src/endpoint/stock-quotes.ts
@@ -1,8 +1,10 @@
 import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
 import { stockEndpointInputParametersDefinition } from '@chainlink/external-adapter-framework/adapter/stock'
+import { TransportRoutes } from '@chainlink/external-adapter-framework/transports'
 import { config } from '../config'
 import overrides from '../config/overrides.json'
-import { transport } from '../transport/stock-quotes'
+import { httpTransport } from '../transport/stock-quotes-http'
+import { wsTransport } from '../transport/stock-quotes-ws'
 import { stockInputParameters } from './utils'
 
 export type BaseEndpointTypes = {
@@ -22,8 +24,11 @@ export type BaseEndpointTypes = {
 
 export const endpoint = new AdapterEndpoint({
   name: 'stock_quotes',
-  aliases: [],
-  transport,
+  transportRoutes: new TransportRoutes<BaseEndpointTypes>()
+    .register('ws', wsTransport)
+    .register('rest', httpTransport),
+  enableCompositeTransport: true,
+  defaultTransport: 'ws',
   inputParameters: stockInputParameters,
   overrides: overrides.finage,
 })